--- src/main/java/com/example/woodenutilities/common/item/plates/WoodenPlateItem.java
@@ -1,11 +1,0 @@
-package com.example.woodenutilities.common.item.plates;
-
-import com.example.woodenutilities.common.init.ModInit;
-import net.minecraft.world.item.Item;
-
-public class WoodenPlateItem extends Item {
-
-    public WoodenPlateItem() {
-        super(new Properties().tab(ModInit.CREATIVE_TAB));
-    }
-}
