--- src/main/java/com/example/woodenutilities/common/utility/ModConstants.java
@@ -1,28 +1,0 @@
-package com.example.woodenutilities.common.utility;
-
-public class ModConstants {
-
-    public static final String MOD_ID = "woodenutilities";
-
-    public static class Items {
-
-        public static final String OAK_PLATE = "oak_plate";
-        public static final String BIRCH_PLATE = "birch_plate";
-        public static final String JUNGLE_PLATE = "jungle_plate";
-        public static final String DARK_OAK_PLATE = "dark_oak_plate";
-        public static final String ACACIA_PLATE = "acacia_plate";
-        public static final String SPRUCE_PLATE = "spruce_plate";
-        public static final String CRIMSON_PLATE = "crimson_plate";
-        public static final String WARPED_PLATE = "warped_plate";
-
-        public static final String WOODEN_SHEARS = "wooden_shears";
-        public static final String WOODEN_BUCKET = "wooden_bucket";
-        public static final String WOODEN_BUCKET_LAVA = "lava_wooden_bucket";
-        public static final String WOODEN_BUCKET_WATER = "water_wooden_bucket";
-        public static final String WOODEN_DIAMOND = "wooden_diamond";
-    }
-
-    public class Blocks {
-        public static final String CRAFTING_SLAB = "crafting_slab";
-    }
-}
