+++ src/main/java/io/github/moderngamingworlds_mods/mods/woodenutilities/common/item/plates/WoodenPlateItem.java
@@ -1,0 +1,11 @@
+package io.github.moderngamingworlds_mods.mods.woodenutilities.common.item.plates;
+
+import io.github.moderngamingworlds_mods.mods.woodenutilities.common.init.ModInit;
+import net.minecraft.world.item.Item;
+
+public class WoodenPlateItem extends Item {
+
+    public WoodenPlateItem() {
+        super(new Properties().tab(ModInit.CREATIVE_TAB));
+    }
+}
